Implement ec_verify_hash and ec_sign_hash
This patch adds support for the ec_verify_hash and ec_sign_hash
functions. These functions takes an additional hash-function parameter
which is used to hash the given input data. This allows us to use
cryptographically strong hash functions to hash input data of arbitrary
length without having to worry about being truncated.

We do not currently emit an error if the given input data is going to be
truncated. This will happen if size_in_bits(InputData) >
ec_curve_size(Curve). This will be fixed in another patch.